support &section=<filename> to override textual part
Some files of the split versions of the textual transcripts may not be
accessable via the by-page index. Thus, the documentViewer allows an
optional parameter &section that bears the plain file name that should
be loaded. If this parameter is present, the corresponding page is
loaded.

Part of the fix for faustedition/faust-gen-html/#15