Highlight fragments (typically verses).
When following a link that contains a fragment (e.g., pointing to a
verse like in
http://beta.faustedition.net/documentViewer.php?faustUri=faust://xml/document/faust/2.3/gsa_391347.xml&page=27&view=text#l8948
), we will now scroll this into view and highlight it even in the
document viewer (by adding the `target` class, the :target selector
alone does not work here).

Currently this is visualized by a gray background and an
indicator-colored and always visible verse number; however, background
color is already taken for the variant apparatus, so suggestions are
welcome.