Start searches in background when already on search page.
Entering a search into the search bar usually triggers the following
steps:

1. The query is sent to the eXist app’s /query endpoint
2. This will redirect either to the shortcut target if the query is a
   shortcut or to the static /search page, with the query string set
3. The search page will then query /search/{text,meta,testimony,info} in
   parallel in the background and display the results when they are ready.

With this patch, we will skip step 1 and 2 when we're already on the
search page. Instead, we run /query in the background and perform the
redirect manually when it’s s shortcut (i.e. no redirection to /search).

This should increase re-query speed, stay on the selected tab (and thus
fixes #584) and reduces the times the user just sees a loading
indicator.